backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
PyMySQL==1.1.0
cryptography==41.0.4